[refactor] Rename MMwithBytes and LLMFusedKernelsTest classes to TestMatMulWithByteArrays and TestRMSNormLayer, respectively, and update package structure
